Move test package to internal
