+++ grpc/internal/test/cancel.go
@@ -1,7 +1,7 @@
 // Copyright © 2017 The Things Network
 // Use of this source code is governed by the MIT license that can be found in the LICENSE file.
 
-package restartstreamtest
+package test
 
 import (
 	"golang.org/x/net/context"
